+++ src/pwm.rs
@@ -25,25 +25,25 @@ pub trait Pins<TIM, P> {
 
     fn split() -> Self::Channels;
 }
-pub use crate::timer::{CPin, C1, C2, C3, C4};
+pub use crate::timer::{CPin, Ch, C1, C2, C3, C4};
 
-pub struct PwmChannel<TIM, CHANNEL> {
-    _channel: PhantomData<CHANNEL>,
-    _tim: PhantomData<TIM>,
+pub struct PwmChannel<TIM, const C: u8> {
+    pub(super) _tim: PhantomData<TIM>,
 }
 
 macro_rules! pins_impl {
-    ( $( ( $($PINX:ident),+ ), ( $($ENCHX:ident),* ); )+ ) => {
+    ( $( ( $($PINX:ident),+ ), ( $($ENCHX:ident),+ ); )+ ) => {
         $(
             #[allow(unused_parens)]
-            impl<TIM, $($PINX,)+> Pins<TIM, ($($ENCHX),+)> for ($($PINX),+)
+            impl<TIM, $($PINX,)+> Pins<TIM, ($(Ch<$ENCHX>),+)> for ($($PINX),+)
             where
-                $($PINX: CPin<$ENCHX, TIM>,)+
+                TIM: WithPwm,
+                $($PINX: CPin<TIM, $ENCHX>,)+
             {
                 $(const $ENCHX: bool = true;)+
                 type Channels = ($(PwmChannel<TIM, $ENCHX>),+);
                 fn split() -> Self::Channels {
-                    ($(PwmChannel::<TIM, $ENCHX> { _channel: PhantomData, _tim: PhantomData }),+)
+                    ($(PwmChannel::<TIM, $ENCHX>::new()),+)
                 }
             }
         )+
@@ -68,71 +68,83 @@ pins_impl!(
     (P4), (C4);
 );
 
-impl<C, TIM, P1: CPin<C, TIM>, P2: CPin<C, TIM>> CPin<C, TIM> for (P1, P2) {}
-impl<C, TIM, P1: CPin<C, TIM>, P2: CPin<C, TIM>, P3: CPin<C, TIM>> CPin<C, TIM> for (P1, P2, P3) {}
-impl<C, TIM, P1: CPin<C, TIM>, P2: CPin<C, TIM>, P3: CPin<C, TIM>, P4: CPin<C, TIM>> CPin<C, TIM>
-    for (P1, P2, P3, P4)
+impl<TIM, P1, P2, const C: u8> CPin<TIM, C> for (P1, P2)
+where
+    P1: CPin<TIM, C>,
+    P2: CPin<TIM, C>,
+{
+}
+impl<TIM, P1, P2, P3, const C: u8> CPin<TIM, C> for (P1, P2, P3)
+where
+    P1: CPin<TIM, C>,
+    P2: CPin<TIM, C>,
+    P3: CPin<TIM, C>,
+{
+}
+impl<TIM, P1, P2, P3, P4, const C: u8> CPin<TIM, C> for (P1, P2, P3, P4)
+where
+    P1: CPin<TIM, C>,
+    P2: CPin<TIM, C>,
+    P3: CPin<TIM, C>,
+    P4: CPin<TIM, C>,
 {
 }
 
-macro_rules! pwm_pin {
-    ($TIMX:ty, $C:ty, $ccr: ident, $bit:literal) => {
-        impl PwmChannel<$TIMX, $C> {
-            #[inline]
-            pub fn disable(&mut self) {
-                //NOTE(unsafe) atomic write with no side effects
-                unsafe { bb::clear(&(*<$TIMX>::ptr()).ccer, $bit) }
-            }
-
-            #[inline]
-            pub fn enable(&mut self) {
-                //NOTE(unsafe) atomic write with no side effects
-                unsafe { bb::set(&(*<$TIMX>::ptr()).ccer, $bit) }
-            }
-
-            #[inline]
-            pub fn get_duty(&self) -> u16 {
-                //NOTE(unsafe) atomic read with no side effects
-                unsafe { (*<$TIMX>::ptr()).$ccr.read().bits() as u16 }
-            }
-
-            /// If `0` returned means max_duty is 2^16
-            #[inline]
-            pub fn get_max_duty(&self) -> u16 {
-                //NOTE(unsafe) atomic read with no side effects
-                unsafe { ((*<$TIMX>::ptr()).arr.read().bits() as u16).wrapping_add(1) }
-            }
-
-            #[inline]
-            pub fn set_duty(&mut self, duty: u16) {
-                //NOTE(unsafe) atomic write with no side effects
-                unsafe { (*<$TIMX>::ptr()).$ccr.write(|w| w.bits(duty.into())) }
-            }
+impl<TIM: WithPwm, const C: u8> PwmChannel<TIM, C> {
+    pub(crate) fn new() -> Self {
+        Self {
+            _tim: core::marker::PhantomData,
         }
-
-        impl embedded_hal::PwmPin for PwmChannel<$TIMX, $C> {
-            type Duty = u16;
-
-            fn disable(&mut self) {
-                self.disable()
-            }
-            fn enable(&mut self) {
-                self.enable()
-            }
-            fn get_duty(&self) -> Self::Duty {
-                self.get_duty()
-            }
-            fn get_max_duty(&self) -> Self::Duty {
-                self.get_max_duty()
-            }
-            fn set_duty(&mut self, duty: Self::Duty) {
-                self.set_duty(duty)
-            }
-        }
-    };
+    }
 }
 
-pub(crate) use pwm_pin;
+impl<TIM: WithPwm, const C: u8> PwmChannel<TIM, C> {
+    #[inline]
+    pub fn disable(&mut self) {
+        TIM::enable_channel(C, false);
+    }
+
+    #[inline]
+    pub fn enable(&mut self) {
+        TIM::enable_channel(C, true);
+    }
+
+    #[inline]
+    pub fn get_duty(&self) -> u16 {
+        TIM::read_cc_value(C) as u16
+    }
+
+    /// If `0` returned means max_duty is 2^16
+    #[inline]
+    pub fn get_max_duty(&self) -> u16 {
+        (TIM::read_auto_reload() as u16).wrapping_add(1)
+    }
+
+    #[inline]
+    pub fn set_duty(&mut self, duty: u16) {
+        TIM::set_cc_value(C, duty as u32)
+    }
+}
+
+impl<TIM: WithPwm, const C: u8> embedded_hal::PwmPin for PwmChannel<TIM, C> {
+    type Duty = u16;
+
+    fn disable(&mut self) {
+        self.disable()
+    }
+    fn enable(&mut self) {
+        self.enable()
+    }
+    fn get_duty(&self) -> Self::Duty {
+        self.get_duty()
+    }
+    fn get_max_duty(&self) -> Self::Duty {
+        self.get_max_duty()
+    }
+    fn set_duty(&mut self, duty: Self::Duty) {
+        self.set_duty(duty)
+    }
+}
 
 impl<TIM: Instance + WithPwm> Timer<TIM> {
     pub fn pwm<P, PINS, T>(mut self, _pins: PINS, freq: T) -> Pwm<TIM, P, PINS>
@@ -207,33 +219,31 @@ where
     type Time = Hertz;
 
     fn enable(&mut self, channel: Self::Channel) {
-        self.tim.enable_channel(PINS::check_used(channel), true)
+        TIM::enable_channel(PINS::check_used(channel) as u8, true)
     }
 
     fn disable(&mut self, channel: Self::Channel) {
-        self.tim.enable_channel(PINS::check_used(channel), false)
+        TIM::enable_channel(PINS::check_used(channel) as u8, false)
     }
 
     fn get_duty(&self, channel: Self::Channel) -> Self::Duty {
-        let duty: u32 = self.tim.read_cc_value(PINS::check_used(channel)).into();
+        let duty: u32 = TIM::read_cc_value(PINS::check_used(channel) as u8).into();
         duty as u16
     }
 
     fn set_duty(&mut self, channel: Self::Channel, duty: Self::Duty) {
-        self.tim
-            .set_cc_value(PINS::check_used(channel), duty.into())
+        TIM::set_cc_value(PINS::check_used(channel) as u8, duty as u32)
     }
 
     /// If `0` returned means max_duty is 2^16
     fn get_max_duty(&self) -> Self::Duty {
-        let arr: u32 = self.tim.read_auto_reload().into();
-        (arr as u16).wrapping_add(1)
+        (TIM::read_auto_reload() as u16).wrapping_add(1)
     }
 
     fn get_period(&self) -> Self::Time {
         let clk = self.clk;
         let psc = self.tim.read_prescaler() as u32;
-        let arr: u32 = self.tim.read_auto_reload().into();
+        let arr = TIM::read_auto_reload();
 
         // Length in ms of an internal clock pulse
         (clk.0 / (psc * arr)).hz()