+++ src/timer.rs
@@ -11,10 +11,12 @@ use void::Void;
 
 use crate::bb;
 use crate::pac::{self, RCC};
-use crate::pwm::{pwm_pin, PwmChannel};
 
 use crate::rcc::{self, Clocks};
 use crate::time::Hertz;
+
+mod pins;
+pub use pins::*;
 
 /// Timer wrapper
 pub struct Timer<TIM> {
@@ -225,7 +227,7 @@ mod sealed {
         type Width: Into<u32> + From<u16>;
         fn max_auto_reload() -> u32;
         fn set_auto_reload(&mut self, arr: u32) -> Result<(), super::Error>;
-        fn read_auto_reload(&self) -> Self::Width;
+        fn read_auto_reload() -> u32;
         fn enable_preload(&mut self, b: bool);
         fn enable_counter(&mut self);
         fn disable_counter(&mut self);
@@ -244,11 +246,11 @@ mod sealed {
 
     pub trait WithPwm: General {
         const CH_NUMBER: u8;
-        fn read_cc_value(&self, channel: Channel) -> Self::Width;
-        fn set_cc_value(&mut self, channel: Channel, value: Self::Width);
+        fn read_cc_value(channel: u8) -> u32;
+        fn set_cc_value(channel: u8, value: u32);
         fn preload_output_channel_in_mode(&mut self, channel: Channel, mode: Ocm);
         fn start_pwm(&mut self);
-        fn enable_channel(&mut self, channel: Channel, b: bool);
+        fn enable_channel(channel: u8, b: bool);
     }
 }
 pub(crate) use sealed::{General, WithPwm};
@@ -303,8 +305,9 @@ macro_rules! hal {
                     }
                 }
                 #[inline(always)]
-                fn read_auto_reload(&self) -> Self::Width {
-                    self.arr.read().bits() as Self::Width
+                fn read_auto_reload() -> u32 {
+                    let tim = unsafe { &*<$TIM>::ptr() };
+                    tim.arr.read().bits()
                 }
                 #[inline(always)]
                 fn enable_preload(&mut self, b: bool) {
@@ -380,21 +383,23 @@ macro_rules! with_pwm {
             const CH_NUMBER: u8 = 1;
 
             #[inline(always)]
-            fn read_cc_value(&self, channel: Channel) -> Self::Width {
+            fn read_cc_value(channel: u8) -> u32 {
+                let tim = unsafe { &*<$TIM>::ptr() };
                 match channel {
-                    Channel::C1 => {
-                        self.ccr1.read().ccr().bits()
+                    0 => {
+                        tim.ccr1.read().bits()
                     }
                     _ => 0,
                 }
             }
 
             #[inline(always)]
-            fn set_cc_value(&mut self, channel: Channel, value: Self::Width) {
+            fn set_cc_value(channel: u8, value: u32) {
+                let tim = unsafe { &*<$TIM>::ptr() };
                 #[allow(unused_unsafe)]
                 match channel {
-                    Channel::C1 => {
-                        self.ccr1.write(|w| unsafe { w.ccr().bits(value) })
+                    0 => {
+                        tim.ccr1.write(|w| unsafe { w.bits(value) })
                     }
                     _ => {},
                 }
@@ -417,46 +422,46 @@ macro_rules! with_pwm {
             }
 
             #[inline(always)]
-            fn enable_channel(&mut self, channel: Channel, b: bool) {
-                let c = channel as u8;
+            fn enable_channel(c: u8, b: bool) {
+                let tim = unsafe { &*<$TIM>::ptr() };
                 if c < Self::CH_NUMBER {
                     if b {
-                        unsafe { bb::set(&self.ccer, c*4) }
+                        unsafe { bb::set(&tim.ccer, c*4) }
                     } else {
-                        unsafe { bb::clear(&self.ccer, c*4) }
+                        unsafe { bb::clear(&tim.ccer, c*4) }
                     }
                 }
             }
         }
-
-        pwm_pin!($TIM, C1, ccr1, 0);
     };
     ($TIM:ty: CH2) => {
         impl WithPwm for $TIM {
             const CH_NUMBER: u8 = 2;
 
             #[inline(always)]
-            fn read_cc_value(&self, channel: Channel) -> Self::Width {
+            fn read_cc_value(channel: u8) -> u32 {
+                let tim = unsafe { &*<$TIM>::ptr() };
                 match channel {
-                    Channel::C1 => {
-                        self.ccr1.read().ccr().bits()
+                    0 => {
+                        tim.ccr1.read().bits()
                     }
-                    Channel::C2 => {
-                        self.ccr2.read().ccr().bits()
+                    1 => {
+                        tim.ccr2.read().bits()
                     }
                     _ => 0,
                 }
             }
 
             #[inline(always)]
-            fn set_cc_value(&mut self, channel: Channel, value: Self::Width) {
+            fn set_cc_value(channel: u8, value: u32) {
+                let tim = unsafe { &*<$TIM>::ptr() };
                 #[allow(unused_unsafe)]
                 match channel {
-                    Channel::C1 => {
-                        self.ccr1.write(|w| unsafe { w.ccr().bits(value) })
+                    0 => {
+                        tim.ccr1.write(|w| unsafe { w.bits(value) })
                     }
-                    Channel::C2 => {
-                        self.ccr2.write(|w| unsafe { w.ccr().bits(value) })
+                    1 => {
+                        tim.ccr2.write(|w| unsafe { w.bits(value) })
                     }
                     _ => {},
                 }
@@ -483,61 +488,60 @@ macro_rules! with_pwm {
             }
 
             #[inline(always)]
-            fn enable_channel(&mut self, channel: Channel, b: bool) {
-                let c = channel as u8;
+            fn enable_channel(c: u8, b: bool) {
+                let tim = unsafe { &*<$TIM>::ptr() };
                 if c < Self::CH_NUMBER {
                     if b {
-                        unsafe { bb::set(&self.ccer, c*4) }
+                        unsafe { bb::set(&tim.ccer, c*4) }
                     } else {
-                        unsafe { bb::clear(&self.ccer, c*4) }
+                        unsafe { bb::clear(&tim.ccer, c*4) }
                     }
                 }
             }
         }
-
-        pwm_pin!($TIM, C1, ccr1, 0);
-        pwm_pin!($TIM, C2, ccr2, 4);
     };
     ($TIM:ty: CH4 $(, $aoe:ident)?) => {
         impl WithPwm for $TIM {
             const CH_NUMBER: u8 = 4;
 
             #[inline(always)]
-            fn read_cc_value(&self, channel: Channel) -> Self::Width {
+            fn read_cc_value(channel: u8) -> u32 {
+                let tim = unsafe { &*<$TIM>::ptr() };
                 let ccr = match channel {
-                    Channel::C1 => {
-                        &self.ccr1
+                    0 => {
+                        &tim.ccr1
                     }
-                    Channel::C2 => {
-                        &self.ccr2
+                    1 => {
+                        &tim.ccr2
                     }
-                    Channel::C3 => {
-                        &self.ccr3
+                    2 => {
+                        &tim.ccr3
                     }
-                    Channel::C4 => {
-                        &self.ccr4
+                    _ => {
+                        &tim.ccr4
                     }
                 };
-                ccr.read().bits() as Self::Width
+                ccr.read().bits()
             }
 
             #[inline(always)]
-            fn set_cc_value(&mut self, channel: Channel, value: Self::Width) {
+            fn set_cc_value(channel: u8, value: u32) {
+                let tim = unsafe { &*<$TIM>::ptr() };
                 let ccr = match channel {
-                    Channel::C1 => {
-                        &self.ccr1
+                    0 => {
+                        &tim.ccr1
                     }
-                    Channel::C2 => {
-                        &self.ccr2
+                    1 => {
+                        &tim.ccr2
                     }
-                    Channel::C3 => {
-                        &self.ccr3
+                    2 => {
+                        &tim.ccr3
                     }
-                    Channel::C4 => {
-                        &self.ccr4
+                    _ => {
+                        &tim.ccr4
                     }
                 };
-                ccr.write(|w| unsafe { w.bits(value as u32) })
+                ccr.write(|w| unsafe { w.bits(value) })
             }
 
             #[inline(always)]
@@ -569,22 +573,17 @@ macro_rules! with_pwm {
             }
 
             #[inline(always)]
-            fn enable_channel(&mut self, channel: Channel, b: bool) {
-                let c = channel as u8;
+            fn enable_channel(c: u8, b: bool) {
+                let tim = unsafe { &*<$TIM>::ptr() };
                 if c < Self::CH_NUMBER {
                     if b {
-                        unsafe { bb::set(&self.ccer, c*4) }
+                        unsafe { bb::set(&tim.ccer, c*4) }
                     } else {
-                        unsafe { bb::clear(&self.ccer, c*4) }
+                        unsafe { bb::clear(&tim.ccer, c*4) }
                     }
                 }
             }
         }
-
-        pwm_pin!($TIM, C1, ccr1, 0);
-        pwm_pin!($TIM, C2, ccr2, 4);
-        pwm_pin!($TIM, C3, ccr3, 8);
-        pwm_pin!($TIM, C4, ccr4, 12);
     }
 }
 
@@ -720,224 +719,4 @@ hal!(
     pac::TIM12: [Timer12, u16, CH2],
     pac::TIM13: [Timer13, u16, CH1],
     pac::TIM14: [Timer14, u16, CH1],
-);
-
-use crate::gpio::{self, Alternate};
-
-// Output channels markers
-pub trait CPin<C, TIM> {}
-pub struct C1;
-pub struct C2;
-pub struct C3;
-pub struct C4;
-
-macro_rules! channel_impl {
-    ( $( $TIM:ident, $C:ident, $PINX:ident, $AF:literal; )+ ) => {
-        $(
-            impl<Otype> CPin<$C, crate::pac::$TIM> for gpio::$PINX<Alternate<Otype, $AF>> { }
-        )+
-    };
-}
-
-// The approach to PWM channel implementation is to group parts with
-// common pins, starting with groupings of the largest number of parts
-// and moving to smaller and smaller groupings.  Last, we have individual
-// parts to cover exceptions.
-
-// All parts have these PWM pins.
-channel_impl!(
-    TIM1, C1, PA8, 1;
-    TIM1, C2, PA9, 1;
-    TIM1, C3, PA10, 1;
-    TIM1, C4, PA11, 1;
-
-    TIM5, C1, PA0, 2;
-    TIM5, C2, PA1, 2;
-    TIM5, C3, PA2, 2;
-    TIM5, C4, PA3, 2;
-
-    TIM9, C1, PA2, 3;
-    TIM9, C2, PA3, 3;
-
-    TIM11, C1, PB9, 3;
-);
-
-// All parts except F410.
-#[cfg(any(
-    feature = "stm32f401",
-    feature = "stm32f405",
-    feature = "stm32f407",
-    feature = "stm32f411",
-    feature = "stm32f412",
-    feature = "stm32f413",
-    feature = "stm32f415",
-    feature = "stm32f417",
-    feature = "stm32f423",
-    feature = "stm32f427",
-    feature = "stm32f429",
-    feature = "stm32f437",
-    feature = "stm32f439",
-    feature = "stm32f446",
-    feature = "stm32f469",
-    feature = "stm32f479"
-))]
-channel_impl!(
-    TIM1, C1, PE9, 1;
-    TIM1, C2, PE11, 1;
-    TIM1, C3, PE13, 1;
-    TIM1, C4, PE14, 1;
-
-    TIM2, C1, PA0, 1;
-    TIM2, C2, PA1, 1;
-    TIM2, C3, PA2, 1;
-    TIM2, C4, PA3, 1;
-
-    TIM2, C2, PB3, 1;
-    TIM2, C3, PB10, 1;
-    TIM2, C4, PB11, 1;
-
-    TIM2, C1, PA5, 1;
-    TIM2, C1, PA15, 1;
-
-    TIM3, C1, PA6, 2;
-    TIM3, C2, PA7, 2;
-    TIM3, C3, PB0, 2;
-    TIM3, C4, PB1, 2;
-
-    TIM3, C1, PB4, 2;
-    TIM3, C2, PB5, 2;
-
-    TIM3, C1, PC6, 2;
-    TIM3, C2, PC7, 2;
-    TIM3, C3, PC8, 2;
-    TIM3, C4, PC9, 2;
-
-    TIM4, C1, PB6, 2;
-    TIM4, C2, PB7, 2;
-    TIM4, C3, PB8, 2;
-    TIM4, C4, PB9, 2;
-
-    TIM4, C1, PD12, 2;
-    TIM4, C2, PD13, 2;
-    TIM4, C3, PD14, 2;
-    TIM4, C4, PD15, 2;
-
-    TIM10, C1, PB8, 3;
-);
-
-// All parts except F401 and F410.
-#[cfg(any(
-    feature = "stm32f405",
-    feature = "stm32f407",
-    feature = "stm32f411",
-    feature = "stm32f412",
-    feature = "stm32f413",
-    feature = "stm32f415",
-    feature = "stm32f417",
-    feature = "stm32f423",
-    feature = "stm32f427",
-    feature = "stm32f429",
-    feature = "stm32f437",
-    feature = "stm32f439",
-    feature = "stm32f446",
-    feature = "stm32f469",
-    feature = "stm32f479"
-))]
-channel_impl!(
-    TIM9, C1, PE5, 3;
-    TIM9, C2, PE6, 3;
-);
-
-// All parts except F401, F410, and F411.
-#[cfg(any(
-    feature = "stm32f405",
-    feature = "stm32f407",
-    feature = "stm32f412",
-    feature = "stm32f413",
-    feature = "stm32f415",
-    feature = "stm32f417",
-    feature = "stm32f423",
-    feature = "stm32f427",
-    feature = "stm32f429",
-    feature = "stm32f437",
-    feature = "stm32f439",
-    feature = "stm32f446",
-    feature = "stm32f469",
-    feature = "stm32f479"
-))]
-channel_impl!(
-    TIM8, C1, PC6, 3;
-    TIM8, C2, PC7, 3;
-    TIM8, C3, PC8, 3;
-    TIM8, C4, PC9, 3;
-
-    TIM10, C1, PF6, 3;
-
-    TIM11, C1, PF7, 3;
-
-    TIM12, C1, PB14, 9;
-    TIM12, C2, PB15, 9;
-
-    TIM13, C1, PA6, 9;
-    TIM13, C1, PF8, 9;  // Not a mistake: TIM13 has only one channel.
-
-    TIM14, C1, PA7, 9;
-    TIM14, C1, PF9, 9;  // Not a mistake: TIM14 has only one channel.
-);
-
-// STM's "advanced and foundation" lines except F446.
-#[cfg(any(
-    feature = "stm32f405",
-    feature = "stm32f407",
-    feature = "stm32f415",
-    feature = "stm32f417",
-    feature = "stm32f427",
-    feature = "stm32f429",
-    feature = "stm32f437",
-    feature = "stm32f439",
-    feature = "stm32f469",
-    feature = "stm32f479"
-))]
-channel_impl!(
-    TIM5, C1, PH10, 2;
-    TIM5, C2, PH11, 2;
-    TIM5, C3, PH12, 2;
-    TIM5, C4, PI0, 2;
-
-    TIM8, C1, PI5, 3;
-    TIM8, C2, PI6, 3;
-    TIM8, C3, PI7, 3;
-    TIM8, C4, PI2, 3;
-
-    TIM12, C1, PH6, 9;
-    TIM12, C2, PH9, 9;
-);
-
-#[cfg(any(feature = "stm32f412", feature = "stm32f413", feature = "stm32f423"))]
-channel_impl!(
-    TIM5, C1, PF3, 2;
-    TIM5, C2, PF4, 2;
-    TIM5, C3, PF5, 2;
-    TIM5, C4, PF10, 2;
-);
-
-#[cfg(feature = "stm32f410")]
-channel_impl!(
-    TIM5, C1, PB12, 2;
-    TIM5, C2, PC10, 2;
-    TIM5, C3, PC11, 2;
-    TIM5, C4, PB11, 2;
-
-    TIM9, C1, PC4, 3;
-    TIM9, C2, PC5, 3;
-
-    TIM11, C1, PC13, 3;
-);
-
-#[cfg(feature = "stm32f446")]
-channel_impl!(
-    TIM2, C1, PB8, 1;
-    TIM2, C2, PB9, 1;
-
-    TIM2, C4, PB2, 1;
 );